Apply code style rules (#951)
* nameof

* add stylecop analyzer

* remove and sort usings

* more

* use builtin type alias

* ignore rules

* documentation reordering, fix parameters

* whitespaces

* documentation fixes

* spaces instead of tabs

* whitespaces

* remove multiple blank lines

* more whitespaces

* more spaces

* Braces for multi-line statements should not share line

* remove empty lines

* pattern matching

* rules for docs

* access modifiers + auto properties

* access modifiers

* more spaces

* remove trailing spaces

* access modifiers

* access modifiers and line separations

* readonly modifier

* disable analyzer by default

* one more fixes

* cosmetics

+++ Source/EasyNetQ.Tests/ConnectionString/ConnectionStringGrammarTests.cs
@@ -44,11 +44,11 @@ namespace EasyNetQ.Tests.ConnectionString
 
             hosts.Count().Should().Be(3);
             hosts.ElementAt(0).Host.Should().Be("host.one");
-            hosts.ElementAt(0).Port.Should().Be((ushort) 1001);
+            hosts.ElementAt(0).Port.Should().Be(1001);
             hosts.ElementAt(1).Host.Should().Be("host.two");
-            hosts.ElementAt(1).Port.Should().Be((ushort) 1002);
+            hosts.ElementAt(1).Port.Should().Be(1002);
             hosts.ElementAt(2).Host.Should().Be("host.three");
-            hosts.ElementAt(2).Port.Should().Be((ushort) 1003);
+            hosts.ElementAt(2).Port.Should().Be(1003);
         }
 
         [Fact]
