+++ Source/EasyNetQ.Tests/TimeoutStrategyTest.cs
@@ -7,7 +7,6 @@ namespace EasyNetQ.Tests
     {
     }
 
-
     public class MessageWithoutTimeoutAttribute
     {
     }
@@ -18,14 +17,14 @@ namespace EasyNetQ.Tests
         public void TestWhenMessagetWithAttribute()
         {
             var timeoutStrategy = new TimeoutStrategy(new ConnectionConfiguration { Timeout = 10 });
-            Assert.Equal((ulong)90, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithTimeoutAttribute)));
+            Assert.Equal(90UL, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithTimeoutAttribute)));
         }
 
         [Fact]
         public void TestWhenPersistentMessagesIsFalse()
         {
             var timeoutStrategy = new TimeoutStrategy(new ConnectionConfiguration { Timeout = 10 });
-            Assert.Equal((ulong)10, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithoutTimeoutAttribute)));
+            Assert.Equal(10UL, timeoutStrategy.GetTimeoutSeconds(typeof(MessageWithoutTimeoutAttribute)));
         }
     }
 }
